feat: revive and modernize core logic (WIP branch since 2024 :D)
This commit includes a large set of changes aimed at improving or restoring the
usefulness of the project after a long hiatus. Main changes (to the best of my recollection):

- handle "new" statx call which is being used by e.g. GNU ls
- improve coverage of existing calls, now supporting: stat, lstat,
  fstat, fstatat, chown, lchown, fchown, fchownat, chmod
- Fix bug where we were not correctly sending file descriptors between
  processes (and replaced read/write loop on socket with sendmsg)
- General cleanup
- Completely remove the use of syscall(). This was required because of
  OpenBSD making it a requirement to only call syscalls through their
  libc functions (
https://marc.info/?l=openbsd-tech&m=169841790407370&w=2 ) but it also
is just a lot simpler and means there is no longer limit of only
supporting amd64 because of hardcoded syscall numbers.
- Add tests to ensure we are covering listed calls

NOTE: This commit revives work started in 2024 and contains accumulated
work-in-progress changes. Some areas may need cleanup or review,
but I am hoping if anyone stumbles on this it is more useful with these
changes pushed than without!